backend/requirements.txt
# ===========================================
# TikTrend Finder - API Requirements
# ===========================================

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
gunicorn==21.2.0
python-multipart>=0.0.6

# Database
databases[postgresql]==0.8.0
asyncpg==0.29.0
sqlalchemy>=1.4.42,<1.5

# Redis
redis>=5.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# OpenAI
openai>=1.10.0

# Validation
pydantic>=2.5.2,<3.0.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Utilities
python-dotenv>=1.0.0
orjson>=3.9.10

# Monitoring
prometheus-client>=0.19.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# ===========================================
# Social Media Automation (Vendor Modules)
# ===========================================
instagrapi>=2.0.0
selenium>=4.15.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
